Add comments to EntityData. Move EntityData.kt, PropertyInfo.kt, and TypeConverterInfo.kt to util directory.
