+++ androidutilities-room/src/main/java/com/heyzeusv/androidutilities/room/RoomProcessor.kt
@@ -11,6 +11,7 @@ import com.google.devtools.ksp.validate
 import com.heyzeusv.androidutilities.room.creators.CsvConverterCreator
 import com.heyzeusv.androidutilities.room.creators.EntityFilesCreator
 import com.heyzeusv.androidutilities.room.creators.RoomDataFileCreator
+import com.heyzeusv.androidutilities.room.util.TypeConverterInfo
 
 class RoomProcessor(
     private val codeGenerator: CodeGenerator,
