+++ androidutilities-room/src/main/java/com/heyzeusv/androidutilities/room/creators/EntityFilesCreator.kt
@@ -5,12 +5,12 @@ import com.google.devtools.ksp.processing.Dependencies
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSClassDeclaration
-import com.heyzeusv.androidutilities.room.CloseClass
-import com.heyzeusv.androidutilities.room.EmbeddedInfo
-import com.heyzeusv.androidutilities.room.EntityData
-import com.heyzeusv.androidutilities.room.FieldInfo
-import com.heyzeusv.androidutilities.room.PropertyInfo
-import com.heyzeusv.androidutilities.room.TypeConverterInfo
+import com.heyzeusv.androidutilities.room.util.CloseClass
+import com.heyzeusv.androidutilities.room.util.EmbeddedInfo
+import com.heyzeusv.androidutilities.room.util.EntityData
+import com.heyzeusv.androidutilities.room.util.FieldInfo
+import com.heyzeusv.androidutilities.room.util.PropertyInfo
+import com.heyzeusv.androidutilities.room.util.TypeConverterInfo
 import com.heyzeusv.androidutilities.room.util.CsvData
 import com.heyzeusv.androidutilities.room.util.CsvInfo
 import com.heyzeusv.androidutilities.room.util.addIndented
